refactor(agent): extract agent_parallel.c to stay under the 2000-line cap
The P0 engine commit pushed server/agent_runtime.c from exactly 2000 (the
hard line-check limit) to 2068. Move the parallel fan-out machinery
(parallel_worker, agent_run_parallel) into a new server/agent_parallel.c
(registered in Makefile AGENT_SRCS + CMakeLists), exposing only agent_run_named
(it stays in agent_runtime.c beside the static helpers it uses). agent_runtime.c
is back to 1960. Also stub handle_delegate_aggregate in test_server_dispatch.c,
which links the dispatch table but not server_compute.o (the new dispatch entry
otherwise leaves an undefined reference). Server + thin client link clean;
unit-test-server-dispatch and unit-test-delegate-ensemble pass.
